+++ src/Console/RestoreCommand.php
@@ -7,12 +7,14 @@ namespace Modsx\Console;
 use Illuminate\Console\Command;
 use Modsx\BackupManager;
 use Modsx\BackupRepository;
+use Modsx\Console\Concerns\ConfirmsDestructiveActions;
 use Modsx\Console\Concerns\InteractsWithModules;
 use Modsx\Exceptions\ModsxException;
 use Modsx\ModuleLocator;
 
 class RestoreCommand extends Command
 {
+    use ConfirmsDestructiveActions;
     use InteractsWithModules;
 
     protected $signature = 'modsx:restore